Rename Vulkan Resource sources
