+++ Modules/Graphics/Core/Sources/Methane/Graphics/Vulkan/CommandListVK.hpp
@@ -27,7 +27,7 @@ Vulkan base template implementation of the command list interface.
 #include "CommandQueueVK.h"
 #include "DeviceVK.h"
 #include "ContextVK.h"
-#include "ResourceVK.h"
+#include "ResourceVK.hpp"
 #include "ProgramBindingsVK.h"
 
 #include <Methane/Graphics/CommandListBase.h>
